refactor: :recycle: 軽微なリファクタ
