+++ lib/core/router/app_router.dart
@@ -17,8 +17,8 @@ import '../../feature/setting/presentation/license_page.dart';
 import '../../feature/setting/presentation/muted_user_list_page.dart';
 import '../../feature/setting/presentation/setting_page.dart';
 import '../../feature/user_list/presentation/following_and_follower_list_page.dart';
+import '../../feature/user_list/presentation/like_user_page.dart';
 import '../../feature/user_profile/presentation/page/edit_profile/edit_profile_page.dart';
-import '../../feature/user_profile/presentation/page/like_user_page.dart/like_user_page.dart';
 import '../../feature/user_profile/presentation/page/profile/profile_page.dart';
 import '../../feature/user_profile/presentation/page/search_user/search_user_page.dart';
 import '../../feature/user_profile/presentation/page/search_user_result/search_user_result_page.dart';
