+++ lib/feature/auth/application/auth_service.dart
@@ -12,10 +12,10 @@ import '../../user_config/repository/device_info_repository.dart';
 import '../../user_config/repository/user_config_repository.dart';
 import '../../user_follow/repository/user_follow_repository.dart';
 import '../../user_list/application/user_id_list_state_notifier.dart';
+import '../../user_list/util/user_list_type.dart';
 import '../../user_profile/domain/user_profile.dart';
 import '../../user_profile/repository/storage_repository.dart';
 import '../../user_profile/repository/user_profile_repository.dart';
-import '../../user_profile/util/user_list_type.dart';
 import '../repository/auth_repository.dart';
 import '../repository/register_user_repository.dart';
 import '../util/constant.dart';
