+++ lib/feature/user_list/presentation/like_user_page.dart
@@ -2,10 +2,10 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../../core/common_widget/simple_widget_for_empty.dart';
-import '../../../../../util/extension/scroll_controller_extension.dart';
-import '../../../../user_list/presentation/profile_list.dart';
-import '../../../util/user_list_type.dart';
+import '../../../core/common_widget/simple_widget_for_empty.dart';
+import '../../../util/extension/scroll_controller_extension.dart';
+import '../util/user_list_type.dart';
+import 'profile_list.dart';
 
 @RoutePage()
 class LikeUserPage extends ConsumerWidget {
