+++ lib/feature/user_list/presentation/profile_list.dart
@@ -7,8 +7,8 @@ import '../../auth/application/auth_state.dart';
 import '../../user_follow/presentation/component/follow_or_unfollow_button.dart';
 import '../../user_profile/presentation/component/profile_tile.dart';
 import '../../user_profile/presentation/component/profile_tile_shimmer.dart';
-import '../../user_profile/util/user_list_type.dart';
 import '../application/user_id_list_state_notifier.dart';
+import '../util/user_list_type.dart';
 
 class ProfileList extends ConsumerWidget {
   const ProfileList({
@@ -25,7 +25,7 @@ class ProfileList extends ConsumerWidget {
   final String? targetDefinitionId;
   final Widget? emptyWidget;
 
-  /// スワイプリフレッシュ時、[userIdListStateNotifierProvider] の 
+  /// スワイプリフレッシュ時、[userIdListStateNotifierProvider] の
   /// invalidate 以外に行う処理。
   final VoidCallback? additionalOnRefresh;
