+++ lib/feature/user_list/repository/fetch_user_list_repository.dart
@@ -5,7 +5,7 @@ import '../../../core/common_provider/firebase_providers.dart';
 import '../../../util/constant/config_constant.dart';
 import '../../../util/constant/firestore_collections.dart';
 import '../../../util/extension/firestore_extension.dart';
-import '../../user_list/domain/user_id_list_state.dart';
+import '../domain/user_id_list_state.dart';
 
 part 'fetch_user_list_repository.g.dart';
